requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
python-dotenv
groq